added

+++ sus/get_next_line/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: otchekai <otchekai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/26 11:33:44 by otchekai          #+#    #+#             */
-/*   Updated: 2022/12/03 18:03:40 by otchekai         ###   ########.fr       */
+/*   Updated: 2022/12/04 13:57:22 by otchekai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,6 +26,6 @@ char	*get_next_line(int fd);
 int		ft_strchr(char *str, int c);
 char	*ft_strjoin(char *s1, char *s2);
 char	*ft_strdup(char *s1);
-int		ft_strlen(char *s);
+size_t		ft_strlen(char *s);
 
 #endif
